[fix] decimal support in currency , brodcast, nsfw cmd in bot
CatUserbot version - 3.0.2

+++ userbot/plugins/nsfw.py
@@ -13,7 +13,7 @@ import requests
 from userbot import catub
 
 from ..helpers.functions import age_verification
-from . import edit_delete, edit_or_reply, reply_id
+from . import edit_delete, edit_or_reply, reply_id, useless
 
 plugin_category = "useless"
 
@@ -33,60 +33,51 @@ plugin_category = "useless"
         ],
     },
 )
-async def danbooru(message):
+async def danbooru(event):
     "Get anime charecter pic or nsfw"
-    reply_to = await reply_id(e)
-    if await age_verification(e, reply_to):
+    reply_to = await reply_id(event)
+    if await age_verification(event, reply_to):
         return
-    await edit_or_reply(message, "`Processing…`")
-    rating = "Explicit" if "nsfw" in message.pattern_match.group(1) else "Safe"
-    search_query = message.pattern_match.group(2)
+    await edit_or_reply(event, "`Processing…`")
+    flag = await useless.importent(event)
+    if flag:
+        return
+    rating = "Explicit" if "nsfw" in event.pattern_match.group(1) else "Safe"
+    search_query = event.pattern_match.group(2)
     params = {
         "limit": 1,
         "random": "true",
         "tags": f"Rating:{rating} {search_query}".strip(),
     }
-
     with requests.get(
         "http://danbooru.donmai.us/posts.json", params=params
     ) as response:
         if response.status_code == 200:
             response = response.json()
         else:
-            await edit_delete(
-                message,
-                f"`An error occurred, response code:` **{response.status_code}**",
-                4,
+            return await edit_delete(
+                event, f"**An error occurred, response code: **`{response.status_code}`"
             )
-            return
 
     if not response:
-        await edit_delete(message, f"`No results for query:` __{search_query}__", 4)
-        return
-
+        return await edit_delete(event, f"**No results for query:** __{search_query}__")
     valid_urls = [
         response[0][url]
         for url in ["file_url", "large_file_url", "source"]
         if url in response[0].keys()
     ]
-
     if not valid_urls:
-        await edit_delete(
-            message, f"`Failed to find URLs for query:` __{search_query}__", 4
+        return await edit_delete(
+            event, f"**Failed to find URLs for query:** __{search_query}__"
         )
-        return
     for image_url in valid_urls:
         try:
-            await message.client.send_file(
-                message.chat_id, image_url, reply_to=reply_to
-            )
-            await message.delete()
+            await event.client.send_file(event.chat_id, image_url, reply_to=reply_to)
+            await event.delete()
             return
         except Exception as e:
-            await edit_or_reply(message, f"{e}")
-    await edit_delete(
-        message, f"``Failed to fetch media for query:` __{search_query}__", 4
-    )
+            await edit_or_reply(event, f"{e}")
+    await edit_delete(event, f"**Failed to fetch media for query:** __{search_query}__")
 
 
 @catub.cat_cmd(
@@ -104,9 +95,11 @@ async def boobs(e):
     if await age_verification(e, reply_to):
         return
     a = await edit_or_reply(e, "`Sending boobs...`")
+    flag = await useless.importent(e)
+    if flag:
+        return
     nsfw = requests.get("http://api.oboobs.ru/noise/1").json()[0]["preview"]
-    urllib.request.urlretrieve("http://media.oboobs.ru/{}".format(nsfw), "*.jpg")
-    os.rename("*.jpg", "boobs.jpg")
+    urllib.request.urlretrieve(f"http://media.oboobs.ru/{nsfw}", "boobs.jpg")
     await e.client.send_file(e.chat_id, "boobs.jpg", reply_to=reply_to)
     os.remove("boobs.jpg")
     await a.delete()
@@ -127,9 +120,11 @@ async def butts(e):
     if await age_verification(e, reply_to):
         return
     a = await edit_or_reply(e, "`Sending beautiful butts...`")
+    flag = await useless.importent(e)
+    if flag:
+        return
     nsfw = requests.get("http://api.obutts.ru/butts/10/1/random").json()[0]["preview"]
-    urllib.request.urlretrieve("http://media.obutts.ru/{}".format(nsfw), "*.jpg")
-    os.rename("*.jpg", "butts.jpg")
+    urllib.request.urlretrieve(f"http://media.obutts.ru/{nsfw}", "butts.jpg")
     await e.client.send_file(e.chat_id, "butts.jpg", reply_to=reply_to)
     os.remove("butts.jpg")
     await a.delete()
